+++ src/main/java/com/prpa/bancodigital/repository/dao/PoliticaTaxaDao.java
@@ -1,6 +1,6 @@
 package com.prpa.bancodigital.repository.dao;
 
-import com.prpa.bancodigital.config.respository.QueryResolver;
+import com.prpa.bancodigital.config.repository.QueryResolver;
 import com.prpa.bancodigital.repository.dao.mapper.PoliticaTaxaMapper;
 import com.prpa.bancodigital.model.PoliticaTaxa;
 import com.prpa.bancodigital.model.enums.TipoTaxa;
