+++ src/main/java/com/prpa/bancodigital/repository/dao/PoliticaUsoDao.java
@@ -1,6 +1,6 @@
 package com.prpa.bancodigital.repository.dao;
 
-import com.prpa.bancodigital.config.respository.QueryResolver;
+import com.prpa.bancodigital.config.repository.QueryResolver;
 import com.prpa.bancodigital.model.PoliticaUso;
 import com.prpa.bancodigital.repository.dao.mapper.PoliticaUsoMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
